+++ app/Http/Submission/Handlers/ConfigureHandler.php
@@ -17,9 +17,9 @@ namespace App\Http\Submission\Handlers;
   PURPOSE. See the above copyright notices for more information.
 =========================================================================*/
 
-use App\Http\Submission\Traits\UpdatesSiteInformation;
 use App\Models\Site;
 use App\Models\SiteInformation;
+use App\Services\SiteService;
 use App\Utils\SubmissionUtils;
 use CDash\Collection\BuildCollection;
 use CDash\Collection\SubscriptionBuilderCollection;
@@ -36,8 +36,6 @@ use CDash\Model\Subscriber;
 
 class ConfigureHandler extends AbstractXmlHandler implements ActionableBuildInterface
 {
-    use UpdatesSiteInformation;
-
     private int $StartTimeStamp = 0;
     private int $EndTimeStamp = 0;
     private array $Builds = [];
@@ -115,7 +113,7 @@ class ConfigureHandler extends AbstractXmlHandler implements ActionableBuildInte
                 $this->BuildName = '(empty)';
             }
 
-            $this->updateSiteInfoIfChanged($this->Site, $siteInformation);
+            SiteService::updateSiteInfoIfChanged($this->Site, $siteInformation);
         } elseif ($name === 'SUBPROJECT') {
             $this->SubProjectName = $attributes['NAME'];
             if (!array_key_exists($this->SubProjectName, $this->SubProjects)) {
