+++ tests/Browser/Pages/BuildCommentsPageTest.php
@@ -2,13 +2,13 @@
 
 namespace Tests\Browser\Pages;
 
-use App\Http\Submission\Traits\UpdatesSiteInformation;
 use App\Models\Build;
 use App\Models\Comment;
 use App\Models\Project;
 use App\Models\Site;
 use App\Models\SiteInformation;
 use App\Models\User;
+use App\Services\SiteService;
 use Illuminate\Support\Str;
 use Laravel\Dusk\Browser;
 use Tests\BrowserTestCase;
@@ -21,7 +21,6 @@ class BuildCommentsPageTest extends BrowserTestCase
     use CreatesProjects;
     use CreatesSites;
     use CreatesUsers;
-    use UpdatesSiteInformation;
 
     private Project $project;
     private Site $site;
@@ -36,7 +35,7 @@ class BuildCommentsPageTest extends BrowserTestCase
         $this->user = $this->makeNormalUser();
 
         $this->site = $this->makeSite();
-        $this->updateSiteInfoIfChanged($this->site, new SiteInformation([]));
+        SiteService::updateSiteInfoIfChanged($this->site, new SiteInformation([]));
     }
 
     public function tearDown(): void
