+++ tests/Browser/Pages/BuildDynamicAnalysisIdPageTest.php
@@ -2,12 +2,12 @@
 
 namespace Tests\Browser\Pages;
 
-use App\Http\Submission\Traits\UpdatesSiteInformation;
 use App\Models\Build;
 use App\Models\DynamicAnalysis;
 use App\Models\Project;
 use App\Models\Site;
 use App\Models\SiteInformation;
+use App\Services\SiteService;
 use Illuminate\Support\Str;
 use Laravel\Dusk\Browser;
 use PHPUnit\Framework\Attributes\DataProvider;
@@ -19,7 +19,6 @@ class BuildDynamicAnalysisIdPageTest extends BrowserTestCase
 {
     use CreatesProjects;
     use CreatesSites;
-    use UpdatesSiteInformation;
 
     private Project $project;
     private Site $site;
@@ -31,7 +30,7 @@ class BuildDynamicAnalysisIdPageTest extends BrowserTestCase
         $this->project = $this->makePublicProject();
 
         $this->site = $this->makeSite();
-        $this->updateSiteInfoIfChanged($this->site, new SiteInformation([]));
+        SiteService::updateSiteInfoIfChanged($this->site, new SiteInformation([]));
     }
 
     public function tearDown(): void
