+++ tests/Browser/Pages/BuildNotesPageTest.php
@@ -2,12 +2,12 @@
 
 namespace Tests\Browser\Pages;
 
-use App\Http\Submission\Traits\UpdatesSiteInformation;
 use App\Models\Build;
 use App\Models\Note;
 use App\Models\Project;
 use App\Models\Site;
 use App\Models\SiteInformation;
+use App\Services\SiteService;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
 use Laravel\Dusk\Browser;
@@ -19,7 +19,6 @@ class BuildNotesPageTest extends BrowserTestCase
 {
     use CreatesProjects;
     use CreatesSites;
-    use UpdatesSiteInformation;
 
     private Project $project;
 
@@ -39,7 +38,7 @@ class BuildNotesPageTest extends BrowserTestCase
         $this->project = $this->makePublicProject();
 
         $this->site = $this->makeSite();
-        $this->updateSiteInfoIfChanged($this->site, new SiteInformation([]));
+        SiteService::updateSiteInfoIfChanged($this->site, new SiteInformation([]));
 
         /** @var Build $build */
         $build = $this->project->builds()->create([
