+++ tests/Browser/Pages/BuildTargetsPageTest.php
@@ -3,12 +3,12 @@
 namespace Tests\Browser\Pages;
 
 use App\Enums\TargetType;
-use App\Http\Submission\Traits\UpdatesSiteInformation;
 use App\Models\Build;
 use App\Models\Project;
 use App\Models\Site;
 use App\Models\SiteInformation;
 use App\Models\Target;
+use App\Services\SiteService;
 use Illuminate\Support\Str;
 use Laravel\Dusk\Browser;
 use Tests\BrowserTestCase;
@@ -19,7 +19,6 @@ class BuildTargetsPageTest extends BrowserTestCase
 {
     use CreatesProjects;
     use CreatesSites;
-    use UpdatesSiteInformation;
 
     private Project $project;
 
@@ -34,7 +33,7 @@ class BuildTargetsPageTest extends BrowserTestCase
         $this->project = $this->makePublicProject();
 
         $this->site = $this->makeSite();
-        $this->updateSiteInfoIfChanged($this->site, new SiteInformation([]));
+        SiteService::updateSiteInfoIfChanged($this->site, new SiteInformation([]));
 
         /** @var Build $build */
         $build = $this->project->builds()->create([
